Allow the discovery of `Servers` which are in the powered `On` state.
Currently we are not running the discovery of `Servers` which are
already powered `On`. A new flag (`--enforce-first-boot`) has been introduced
to enforce the discovery of already powered on `Servers`. The default of
this flag is set to `false`.

Additionally if we encounter a `Server` with `.spec.power==""` in the initial
state, we mutate the `Server` spec to macht the detected `PowerState` in
the `Servers` status.

Co-authored-by: Hardik Dodiya <hardik.dodiya@sap.com>