Add `metal` user and SSH Keypair during `Discovery` boot (#219)
* Add `metal` user and SSH Keypair during `Discovery` boot

This PR introduces a feature to create a `metal` user with corresponding
SSH keypairs which are stored as a `Secret` alongside the discovery
`ServerBootConfiguration`. The `Secret` is named
`<serverbootconfig.name-ssh>`. This allows an administrator to use the
private key in this `Secret` to SSH into the `Server` in case something
went wrong when running the `metalprobe` agent.

* Fix linting issues and bump tools versions

* Add brypter user password hash

* Addressed review comments