Add `MaintenanceBootConfigurationRef` to `Server` type
The `MaintenanceBootConfigurationRef` is used to boot a `Server` into a
maintenance mode. This field is superseeding the `BootConfigurationRef`
which is used for a regular/discovery boot. The reason to introduce a
new field here is, that in case of a `ServerClaim` claiming a `Server`
the `ServerBootConfigurationRef` field is set with a reference to a
`ServerBootConfiguration` which is generated out of a `ServerClaim`.