Option for naming `Servers` as the `BMC` objects (#362)
While the redfish standard supports an API with multiple
systems managed by one endpoint (such as Cisco IMC and others),
an often use-case is to communicate with the BMC directly.

In Netbox, one usually models the system managed by the BMC
and the BMC itself is just an attribute on the system.

The expectation in these circumstances is, that the Server
is not a child of the BMC (i.e. suffix -system-0), but
they are both belonging to the same server (i.e. have the same
name).

The CLI option disable-server-name-suffix allows one to opt-in into the
logic that such a Server should be named as the BMC.