Add flag to facilitate server-name migration
Adding the flag `--disable-server-name-suffix` allows one to remove
the `-system-0` suffix, but provides no path for migration.
It would just create a second server without the suffix.

The flag `--migrate-server-names` supports the migration by deleting
the server only if there is no ServerClaimRef. If there is one, it will
be reported as an error and trigger the exponential backoff.

That way, the BMCs are polled until all Servers have been migrated,
and the flag can be removed.