Handle deleting of server maintenance
When BiosSettings cr is removed a ServerMaintenance cr should also be
removed. Currently it is handled by owner reference but this solution
doesn't work because BiosSettings is a cluster scoped resource and
ServerMaintenance is namespaced resource. This commit introduces
deleting ServerMaintenance when BiosSettings is deleted in a reconcile
function. It also changes owning to watching of ServerMaintenance by
BiosSettings controller.