Update ServerClaimRef in Server instead of Patch
Patch opts out of the optimistic locking functionality of the
kubernetes api, and therefore doesn't protect against
concurrent modifications.

The lock in the controller only protects against concurrent
execution of the same function in the same controller,
but does nothing to stop concurrent modifications
with any other user of the object.